+++ program/source/layout.cpp
@@ -1,7 +1,6 @@
 #include <memory>
 
 #include "glm/ext/vector_double3.hpp"
-#include "glm/ext/vector_int3.hpp"
 
 #include "camera.hpp"
 #include "game.hpp"
@@ -17,19 +16,19 @@ namespace csg
   void scene::main(const std::shared_ptr<scene> &scene)
   {
     scene->set<csg::camera>(glm::dvec3{0.0, 0.0, 80.0})
-      .set<player>("player", glm::ivec3{0.0, 0.0, 0.0})
-      .set<environment>("floor", glm::ivec3{0.0, -61.0, 0.0}, image::floor, animation::floor.main)
-      .set<environment>("shop", glm::ivec3{80.0, 24.0, -1.0}, image::shop, animation::shop.main)
-      .set<environment>("background1", glm::ivec3{0.0, 80.0, -3.0}, image::background1, animation::background1.main)
-      .set<environment>("background2", glm::ivec3{0.0, 80.0, -6.0}, image::background2, animation::background2.main)
-      .set<environment>("background3", glm::ivec3{0.0, 80.0, -9.0}, image::background3, animation::background3.main);
+      .set<player>("player", glm::dvec3{0.0, 0.0, 0.0})
+      .set<environment>("floor", glm::dvec3{0.0, -61.0, 0.0}, image::floor, animation::floor.main)
+      .set<environment>("shop", glm::dvec3{80.0, 24.0, -1.0}, image::shop, animation::shop.main)
+      .set<environment>("background1", glm::dvec3{0.0, 80.0, -3.0}, image::background1, animation::background1.main)
+      .set<environment>("background2", glm::dvec3{0.0, 80.0, -6.0}, image::background2, animation::background2.main)
+      .set<environment>("background3", glm::dvec3{0.0, 80.0, -9.0}, image::background3, animation::background3.main);
   }
 
   void scene::other(const std::shared_ptr<scene> &scene)
   {
     scene->set<csg::camera>(glm::dvec3{0.0, 0.0, 80.0})
-      .set<player>("player", glm::ivec3{0.0, 0.0, 0.0})
-      .set<environment>("floor", glm::ivec3{0.0, -61.0, 0.0}, image::floor, animation::floor.main)
-      .set<environment>("background3", glm::ivec3{0.0, 80.0, -9.0}, image::background3, animation::background3.main);
+      .set<player>("player", glm::dvec3{0.0, 0.0, 0.0})
+      .set<environment>("floor", glm::dvec3{0.0, -61.0, 0.0}, image::floor, animation::floor.main)
+      .set<environment>("background3", glm::dvec3{0.0, 80.0, -9.0}, image::background3, animation::background3.main);
   }
 }
